refactor(cli): extract cli.rs and cmd_daemon.rs from main.rs (#582)
Split main.rs (897 lines) into three focused files:
- cli.rs: Cmd enum + Args struct (351 lines)
- cmd_daemon.rs: run_daemon function (75 lines)
- main.rs: dispatcher + run_join (482 lines)

No logic changes — pure code reorganization.
